refactored structure for different stabilizer chain builders
