+++ src/group/stabchain/mod.rs
@@ -1,13 +1,13 @@
 pub mod element_testing;
 mod moved_point_selector;
+mod stabchain_builder_naive;
 
 use crate::group::orbit::transversal::Transversal;
 use crate::group::Group;
 use crate::perm::Permutation;
 use moved_point_selector::MovedPointSelector;
 
-use std::collections::{HashMap, VecDeque};
-use std::iter::FromIterator;
+use std::collections::HashMap;
 
 #[derive(Debug)]
 pub struct Stabchain {
@@ -16,7 +16,7 @@ pub struct Stabchain {
 
 impl Stabchain {
     fn new_impl(g: &Group, selector: impl MovedPointSelector) -> Self {
-        let mut builder = StabchainBuilder::new(selector);
+        let mut builder = stabchain_builder_naive::StabchainBuilder::new(selector);
         for gen in g.generators() {
             builder.extend(gen.clone());
         }
@@ -106,151 +106,6 @@ impl IntoIterator for Stabchain {
 
     fn into_iter(self) -> Self::IntoIter {
         self.chain.into_iter()
-    }
-}
-
-// Helper struct, used to build the stabilizer chain
-struct StabchainBuilder<T: MovedPointSelector> {
-    current_pos: usize,
-    chain: Vec<StabchainRecord>,
-    selector: T,
-}
-
-impl<T: MovedPointSelector> StabchainBuilder<T> {
-    fn new(selector: T) -> Self {
-        StabchainBuilder {
-            current_pos: 0,
-            chain: Vec::new(),
-            selector,
-        }
-    }
-
-    fn bottom_of_the_chain(&self) -> bool {
-        self.current_pos == self.chain.len()
-    }
-
-    fn current_chain(&self) -> impl Iterator<Item = &StabchainRecord> {
-        self.chain.iter().skip(self.current_pos)
-    }
-
-    fn extend_lower_level(&mut self, p: Permutation) {
-        self.current_pos += 1;
-        self.extend_inner(p);
-        self.current_pos -= 1;
-    }
-
-    #[allow(clippy::map_entry)]
-    fn extend_inner(&mut self, p: Permutation) {
-        // Note that id always in group
-        if element_testing::is_in_group(self.current_chain(), &p) {
-            return;
-        }
-
-        // Bottom of the chain
-        if self.bottom_of_the_chain() {
-            let moved_point = self.selector.moved_point(&p);
-            let mut record = StabchainRecord {
-                base: moved_point,
-                gens: Group::new(&[p.clone()]),
-                transversal: [(moved_point, Permutation::id())].iter().cloned().collect(),
-            };
-
-            let mut next_orbit_point = p.apply(moved_point);
-            let mut representative = p.clone();
-            while next_orbit_point != moved_point {
-                record
-                    .transversal
-                    .insert(next_orbit_point, representative.clone());
-                next_orbit_point = p.apply(next_orbit_point);
-                representative = representative.multiply(&p);
-            }
-            self.chain.push(record);
-            self.extend_lower_level(representative);
-            return;
-        }
-
-        // Then we already had something in this layer
-        // Gets the record to be updated
-        let mut record = self.chain[self.current_pos].clone();
-
-        let mut to_check = VecDeque::from_iter(record.transversal.keys().copied());
-        let mut new_transversal = HashMap::new();
-        while !to_check.is_empty() {
-            let orbit_element = to_check.pop_back().unwrap();
-            let orbit_element_repr = record.transversal.get(&orbit_element).unwrap();
-            let new_image = p.apply(orbit_element);
-
-            // If we already saw the element
-            if record.transversal.contains_key(&new_image)
-                || new_transversal.contains_key(&new_image)
-            {
-                let image_repr = record
-                    .transversal
-                    .get(&new_image)
-                    .or_else(|| new_transversal.get(&new_image))
-                    .unwrap();
-
-                let new_perm = orbit_element_repr.multiply(&p).multiply(&image_repr.inv());
-                self.extend_lower_level(new_perm);
-            } else {
-                new_transversal.insert(new_image, orbit_element_repr.multiply(&p));
-            }
-        }
-
-        // We now want to check all the newly added elements
-        let mut to_check =
-            VecDeque::from_iter(new_transversal.iter().map(|(o, p)| (*o, p.clone())));
-
-        // Update the record
-        record.transversal.extend(new_transversal);
-
-        // While we have orbit elements (and representatives to check)
-        while !to_check.is_empty() {
-            // Get the pair
-            let (orbit_element, orbit_element_repr) = to_check.pop_back().unwrap();
-
-            // For each generator (and p)
-            for generator in std::iter::once(&p).chain(record.gens.generators()) {
-                let new_image = generator.apply(orbit_element);
-
-                // If we have already seen the image
-                if record.transversal.contains_key(&new_image) {
-                    // Get the representative
-                    let image_repr = record.transversal.get(&new_image).unwrap();
-
-                    // Extend lower level
-                    let new_perm = orbit_element_repr
-                        .multiply(generator)
-                        .multiply(&image_repr.inv());
-                    self.extend_lower_level(new_perm);
-                } else {
-                    // Compute the repr s.t. repr^(orbit_element_repr * generator) = orbit_element ^ generator = new_image
-                    let repr = orbit_element_repr.multiply(generator);
-
-                    // Store in transversal
-                    record.transversal.insert(new_image, repr.clone());
-
-                    // Update and ask to check the new image
-                    to_check.push_back((new_image, repr));
-                }
-            }
-        }
-
-        // Update the generators adding p
-        record.gens =
-            Group::from_iter(std::iter::once(&p).chain(record.gens.generators()).cloned());
-
-        // Store the updated record in the chain
-        self.chain[self.current_pos] = record;
-    }
-
-    fn extend(&mut self, p: Permutation) {
-        self.current_pos = 0;
-        self.extend_inner(p);
-    }
-
-    fn build(self) -> Stabchain {
-        Stabchain { chain: self.chain }
     }
 }
